splitting to internal/dotformatting
